% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_ls_from_gis}
\alias{get_ls_from_gis}
\title{Retrieve lift station data from GIS}
\usage{
get_ls_from_gis()
}
\value{
A data frame with columns: \code{STREET_NAME}, \code{STREET_NUMBER},
\code{CMMS_ASSET_ID}, \code{PI_TAG}.
}
\description{
Downloads pump station features from GIS via gisimport package,
then filters#' to in-service JEA lift stations with an CMMS asset ID
}
\details{
Filters applied:
\itemize{
\item \code{FAC_STATE == "INS"} (in-service only)
\item \code{CMMS_ASSET_ID} is not \code{NA}
\item \code{SUBTYPE == 1} (lift stations)
}
}
\keyword{internal}
