% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{safe_get_webid}
\alias{safe_get_webid}
\title{Safe wrapper around \code{\link[piwebapi:get_webid]{piwebapi::get_webid()}}}
\usage{
safe_get_webid(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[piwebapi:get_webid]{piwebapi::get_webid()}}.}
}
\value{
A web ID string, or \code{NULL} if the tag does not exist.
}
\description{
Returns \code{NULL} instead of raising an error when a tag is not found.
(https://purrr.tidyverse.org/reference/possibly.html)
}
\keyword{internal}
